backend/requirements.txt
Django==4.2.1
django-crispy-forms==2.0
crispy-tailwind==0.5.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
markdown==3.4.3
django-filter==23.2
Pillow==10.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis>=4.5.5
celery>=5.3.0
django-environ>=0.10.0
django-cors-headers>=4.1.0
pytest>=7.3.1
pytest-django>=4.5.2
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0 